+++ Backend/src/ConsertaPraMim.Application/Interfaces/IAdminDashboardService.cs
@@ -1,0 +1,8 @@
+using ConsertaPraMim.Application.DTOs;
+
+namespace ConsertaPraMim.Application.Interfaces;
+
+public interface IAdminDashboardService
+{
+    Task<AdminDashboardDto> GetDashboardAsync(AdminDashboardQueryDto query);
+}